requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-registration-redux==2.9
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
future==0.18.2
Pillow==8.2.0
pip==21.1.2
pytz==2021.1
setuptools==56.0.0
six==1.16.0
sqlparse==0.4.1
